+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - [#166][is_166]: Start using Postmark for emails ([#168][pr_168])
 - [#167][is_167]: Create and schedule weekly report email ([#169][pr_169])
 
+### Development
+
+- Regenerate IDE helper markup on models
+
 [is_166]: https://github.com/JSn1nj4/ElliotDerhay.com/issues/166
 
 [is_167]: https://github.com/JSn1nj4/ElliotDerhay.com/issues/167